+++ src/main/clojure/com/github/clojure_repl/intellij/configuration/factory/local.clj
@@ -45,7 +45,6 @@
         [host (parse-long port)])))
 
 (defn ^:private setup-process [^Project project command env-vars]
-  (logger/info "----------__>" env-vars)
   (let [command-str (string/join " " command)
         command-line  (doto (GeneralCommandLine. ^java.util.List command)
                         (.setCharset (Charset/forName "UTF-8"))