+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix entries history navigation in REPL.
 - Fix Local repl configuration setup when more than one project is opened.
+- Fix default shortcuts being added for already customized shortcuts. #94
 
 ## 1.5.0